+++ src/levenshtein.cpp
@@ -15,7 +15,7 @@
  * @param  z третье число
  * @return   самое минимальное значение из 3-х чисел
  */
-const float anyks::LEV::minimum(const float x, const float y, const float z) const noexcept {
+const double anyks::LEV::minimum(const double x, const double y, const double z) const noexcept {
 	// Определяем минимальное значение
 	return lvmin(lvmin(x, y), z);
 }
@@ -70,13 +70,13 @@ const u_short anyks::LEV::mulct(const wstring & pattern, const wstring & text) c
 		 * @param a символ вставки
 		 * @return  вес символа
 		 */
-		weighter.ins = [](const wchar_t a) noexcept {return 2.0f;};
+		weighter.ins = [](const wchar_t a) noexcept {return 2.0;};
 		/**
 		 * Функция для определения веса при удалении
 		 * @param a символ удаления
 		 * @return  вес символа
 		 */
-		weighter.del = [](const wchar_t a) noexcept {return 2.0f;};
+		weighter.del = [](const wchar_t a) noexcept {return 2.0;};
 		/**
 		 * Функция для определения веса при замене
 		 * @param a заменяемый символ
@@ -84,7 +84,7 @@ const u_short anyks::LEV::mulct(const wstring & pattern, const wstring & text) c
 		 * @return  вес символа
 		 */
 		weighter.rep = [](const wchar_t a, const wchar_t b, const wchar_t c, const wchar_t d) noexcept {
-			return ((a == d) && (b == c) ? 0.0f : 1.0f);
+			return ((a == d) && (b == c) ? 0.0 : 1.0);
 		};
 		// Выполняем сравнение
 		result = u_short(this->weighted(pattern, text, &weighter));
@@ -132,13 +132,13 @@ const u_short anyks::LEV::damerau(const wstring & pattern, const wstring & text)
 		 * @param a символ вставки
 		 * @return  вес символа
 		 */
-		weighter.ins = [](const wchar_t a) noexcept {return 1.0f;};
+		weighter.ins = [](const wchar_t a) noexcept {return 1.0;};
 		/**
 		 * Функция для определения веса при удалении
 		 * @param a символ удаления
 		 * @return  вес символа
 		 */
-		weighter.del = [](const wchar_t a) noexcept {return 1.0f;};
+		weighter.del = [](const wchar_t a) noexcept {return 1.0;};
 		/**
 		 * Функция для определения веса при замене
 		 * @param a заменяемый символ
@@ -146,7 +146,7 @@ const u_short anyks::LEV::damerau(const wstring & pattern, const wstring & text)
 		 * @return  вес символа
 		 */
 		weighter.rep = [](const wchar_t a, const wchar_t b, const wchar_t c, const wchar_t d) noexcept {
-			return 1.0f;
+			return 1.0;
 		};
 		// Выполняем сравнение
 		result = u_short(this->weighted(pattern, text, &weighter));
@@ -235,9 +235,9 @@ const u_short anyks::LEV::distance(const wstring & pattern, const wstring & text
  * @param  stl    размер подстроки при сравнении двух слов (от 1 до минимального размера слова)
  * @return        коэффициент Танимото
  */
-const float anyks::LEV::tanimoto(const string & first, const string & second, const u_short stl) const noexcept {
+const double anyks::LEV::tanimoto(const string & first, const string & second, const u_short stl) const noexcept {
 	// Результат работы функции
-	float result = 0.0f;
+	double result = 0.0;
 	// Если первое и второе слово переданы
 	if(!first.empty() && !second.empty()){
 		// Объявляем конвертер
@@ -258,13 +258,13 @@ const float anyks::LEV::tanimoto(const string & first, const string & second, co
  * @param  stl    размер подстроки при сравнении двух слов (от 1 до минимального размера слова)
  * @return        коэффициент Танимото
  */
-const float anyks::LEV::tanimoto(const wstring & first, const wstring & second, const u_short stl) const noexcept {
+const double anyks::LEV::tanimoto(const wstring & first, const wstring & second, const u_short stl) const noexcept {
 	// Результат работы функции
-	float result = 0.0f;
+	double result = 0.0;
 	// Если первое и второе слово переданы
 	if((first.length() > 1) && (second.length() > 1)){
 		// Количество эквивалентных подстрок
-		float esubCount = 0.0f;
+		double esubCount = 0.0;
 		// Подстроки для расчёта схожести
 		wstring subFirst = L"", subSecond = L"";
 		// Получаем размер подстроки
@@ -290,7 +290,7 @@ const float anyks::LEV::tanimoto(const wstring & first, const wstring & second, 
 						// Запоминаем что подстроки совпадают
 						tokens[j] = true;
 						// Увеличиваем количество эквивалентных подстрок
-						esubCount += 1.0f;
+						esubCount += 1.0;
 						// Выходим из цикла
 						break;
 					}
@@ -298,7 +298,7 @@ const float anyks::LEV::tanimoto(const wstring & first, const wstring & second, 
 			}
 		}
 		// Выполняем расчет коэффициента Танимото
-		result = ((1.0f * esubCount) / ((float(firstToken) + float(secondToken)) - esubCount));
+		result = ((1.0 * esubCount) / ((double(firstToken) + double(secondToken)) - esubCount));
 	}
 	// Выводим результат
 	return result;
@@ -310,7 +310,7 @@ const float anyks::LEV::tanimoto(const wstring & first, const wstring & second, 
  * @param  weighter функции сравнения весов
  * @return          вес сравниваемого слова
  */
-const float anyks::LEV::weighted(const string & first, const string & second, const weighter_t * weighter) const noexcept {
+const double anyks::LEV::weighted(const string & first, const string & second, const weighter_t * weighter) const noexcept {
 	// Объявляем конвертер
 	wstring_convert <codecvt_utf8 <wchar_t>> conv;
 	// Конвертируем строки
@@ -326,7 +326,7 @@ const float anyks::LEV::weighted(const string & first, const string & second, co
  * @param  weighter функции сравнения весов
  * @return          вес сравниваемого слова
  */
-const float anyks::LEV::weighted(const wstring & first, const wstring & second, const weighter_t * weighter) const noexcept {
+const double anyks::LEV::weighted(const wstring & first, const wstring & second, const weighter_t * weighter) const noexcept {
 	/**
 	 * calc Функция расчета весов
 	 * @param  first    первое слово
@@ -336,13 +336,13 @@ const float anyks::LEV::weighted(const wstring & first, const wstring & second, 
 	 */
 	auto calc = [this](const wstring & first, const wstring & second, const weighter_t * weighter = nullptr) noexcept {
 		// Результат работы функции
-		float result = 0.0f;
+		double result = 0.0;
 		// Если данные переданы
 		if(weighter != nullptr){
 			// Буквы для сравнения
 			wchar_t a, b, c, d;
 			// Вспомогательные переменные
-			float ic = 0.0f, dc = 0.0f, rc = 0.0f, last = 0.0f, old = 0.0f;
+			double ic = 0.0, dc = 0.0, rc = 0.0, last = 0.0, old = 0.0;
 			// Получаем размеры входящих слов
 			size_t flen = first.length(), slen = second.length();
 			// Если одно из слов не содержит символов
@@ -355,9 +355,9 @@ const float anyks::LEV::weighted(const wstring & first, const wstring & second, 
 				return result;
 			}
 			// Выделяем память для массива
-			vector <float> column((slen + 1) * sizeof(float));
+			vector <double> column((slen + 1) * sizeof(double));
 			// Зануляем первый элемент
-			column.at(0) = 0.0f;
+			column.at(0) = 0.0;
 			// Заполняем символы в таблице для второго слова
 			for(size_t j = 1; j <= slen; ++j) column.at(j) = (column[j - 1] + weighter->ins(second[j - 1]));
 			// Выполняем обход первого слова
@@ -405,13 +405,13 @@ const float anyks::LEV::weighted(const wstring & first, const wstring & second, 
 	 * @param a символ вставки
 	 * @return  вес символа
 	 */
-	defWeighter.ins = [](const wchar_t a) noexcept {return float(0.6f);};
+	defWeighter.ins = [](const wchar_t a) noexcept {return double(0.6);};
 	/**
 	 * Функция для определения веса при удалении
 	 * @param a символ удаления
 	 * @return  вес символа
 	 */
-	defWeighter.del = [](const wchar_t a) noexcept {return float(0.55f);};
+	defWeighter.del = [](const wchar_t a) noexcept {return double(0.55);};
 	/**
 	 * Функция для определения веса при замене
 	 * @param a заменяемый символ
@@ -419,7 +419,7 @@ const float anyks::LEV::weighted(const wstring & first, const wstring & second, 
 	 * @return  вес символа
 	 */
 	defWeighter.rep = [](const wchar_t a, const wchar_t b, const wchar_t c, const wchar_t d) noexcept {
-		return float((a == d) && (b == c) ? -0.2 : 0.5);
+		return double((a == d) && (b == c) ? -0.2 : 0.5);
 	};
 	// Выполняем расчет
 	return calc(first, second, (weighter != nullptr ? weighter : &defWeighter));
